+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/package-info.java
@@ -20,7 +20,7 @@
  * Package configuration values:
  * <ul>
  * <li>{@link is.codion.swing.common.ui.dialog.ExceptionDialogBuilder#SYSTEM_PROPERTIES}
- * <li>{@link is.codion.swing.common.ui.dialog.LookAndFeelSelectionDialogBuilder#INCLUDE_PLATFORM_LOOK_AND_FEELS}
+ * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelComboBox#INCLUDE_PLATFORM_LOOK_AND_FEELS}
  * </ul>
  */
 package is.codion.swing.common.ui.dialog;