+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/laf/LookAndFeelComboBoxTest.java
@@ -24,6 +24,6 @@ public final class LookAndFeelComboBoxTest {
 
 	@Test
 	void test() {
-		LookAndFeelComboBox.lookAndFeelComboBox(true).selectedLookAndFeel();
+		LookAndFeelComboBox.builder().enableOnSelection(true).build().selectedLookAndFeel();
 	}
 }